added fire pits to the scene
re-enabled the auto exposure for "story" it will act as a fade in once
the person pushes play. but can be turned off if the group thinks its
best. Also added the pottery to the scene. If you update this file from
this update and on please bring a copy on a flashdrive or external so
presentation is easier on you guys, i will be bringing a copy on a flash
drive (this update)